crawler/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
requests==2.31.0
selenium==4.15.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
python-dotenv==1.0.0